backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.39.0
requests==2.32.3
sqlalchemy==2.0.35